hw05: Fix after review 1: change ID types to long from int.

+++ hw05/src/main/java/yaklyushkin_otus/spring_2019_02/hw05/builders/GenreBuilder.java
@@ -18,13 +18,13 @@ public class GenreBuilder {
                 ErrorConsts.ERRMSG_WRONGDATA_GENRE_NAME, null, locale);
     }
 
-    public final Genre build(int genreId, String genreName) throws WrongDataException {
+    public final Genre build(long genreId, String genreName) throws WrongDataException {
         this.checkId(genreId);
         this.checkName(genreName);
         return new Genre(genreId, genreName);
     }
 
-    private void checkId(int genreId) throws WrongDataException {
+    private void checkId(long genreId) throws WrongDataException {
         if (genreId < 0) {
             throw new WrongDataException(String.format(this.msgErrorGenreId, genreId));
         }
